на сраницу с постами добавлена регистрация , вход/выход, в модели постов добавлено поле user, возможность обновлять, удалять посты только авторизованым пользователям,и тем кто их создал

+++ books/views.py
@@ -5,6 +5,7 @@ from django.http import HttpResponse
 
 
 def books(request):
+
     books = Book.objects.all()
     genres = Genre.objects.all()
     publishers = Publisher.objects.all()
@@ -48,36 +49,39 @@ def get_tag_books(request, title):
 
 
 def add_book(request):
-    if request.method == "GET":
-        form = BookForm()
-        return render(request, "add_book.html", context={"form": form})
-    elif request.method == "POST":
-        publisher_id = request.POST['publisher']
-        genre_id = request.POST['genre']
+    if request.user.is_authenticated:
+        if request.method == "GET":
+            form = BookForm()
+            return render(request, "add_book.html", context={"form": form})
+        elif request.method == "POST":
+            publisher_id = request.POST['publisher']
+            genre_id = request.POST['genre']
 
-        publisher = None
-        genre = None
-        image = request.FILES.get('image', 'no_image.png')
+            publisher = None
+            genre = None
+            image = request.FILES.get('image', 'no_image.png')
 
-        if publisher_id != '':
-            publisher = Publisher.objects.get(id=publisher_id)
+            if publisher_id != '':
+                publisher = Publisher.objects.get(id=publisher_id)
 
-        if genre_id != '':
-            genre = Genre.objects.get(id=genre_id)
+            if genre_id != '':
+                genre = Genre.objects.get(id=genre_id)
 
-        book = Book.objects.create(title=request.POST['title'],
-                            author=request.POST['author'],
-                            year=request.POST['year'],
-                            raiting=request.POST['raiting'],
-                            publisher=publisher,
-                            genre=genre,
-                            image=image)
-        tags = request.POST.getlist('tags')
-        book.tags.set(tags)
-        book.save()
+            book = Book.objects.create(title=request.POST['title'],
+                                author=request.POST['author'],
+                                year=request.POST['year'],
+                                raiting=request.POST['raiting'],
+                                publisher=publisher,
+                                genre=genre,
+                                image=image,
+                                user=request.user)
+            tags = request.POST.getlist('tags')
+            book.tags.set(tags)
+            book.save()
 
-        return redirect("books")
-
+            return redirect("books")
+    else:
+        return HttpResponse("<h1>Только авторизованный пользователь может добавить книгу!</h1>")
 
     # if request.method == "GET":
     #     form = BookForm()
@@ -153,8 +157,12 @@ def delete_book(request, id):
         book = Book.objects.get(id=id)
     except Book.DoesNotExist:
         return HttpResponse(f"<h1>Книги с  id {id} не существует</h1>")
-    book.delete()
-    return redirect('books')
+
+    if request.user.username != book.user.username:
+        return HttpResponse("<h1> У вас нет прав на удаление этой книги </h1>")
+    else:
+        book.delete()
+        return redirect('books')
 
 
 
@@ -166,34 +174,40 @@ def update_book(request, id):
     except Book.DoesNotExist:
         return HttpResponse(f"<h1>Книги с  id {id} не существует</h1>")
 
-    if request.method == "GET":
-        form = BookForm(instance=book)
-        return render(request, "update_book.html", context={"form": form,
-                                                            "book": book})
+    if request.user.username != book.user.username:
+        return HttpResponse("<h1> У вас нет прав на обновление этой книги </h1>")
+
     else:
-        publisher_id = request.POST['publisher']
-        genre_id = request.POST['genre']
+        if request.method == "GET":
+            form = BookForm(instance=book)
+            return render(request, "update_book.html", context={"form": form,
+                                                                "book": book})
+        else:
+            publisher_id = request.POST['publisher']
+            genre_id = request.POST['genre']
 
-        publisher = None
-        genre = None
-        image = request.FILES.get('image', 'no_image.png')
+            publisher = None
+            genre = None
+            image = request.FILES.get('image', 'no_image.png')
 
-        if publisher_id != '':
-            publisher = Publisher.objects.get(id=publisher_id)
+            if publisher_id != '':
+                publisher = Publisher.objects.get(id=publisher_id)
 
-        if genre_id != '':
-            genre = Genre.objects.get(id=genre_id)
+            if genre_id != '':
+                genre = Genre.objects.get(id=genre_id)
 
-        book.title = request.POST['title']
-        book.author = request.POST['author']
-        book.year = request.POST['year']
-        book.raiting = request.POST['raiting']
-        book.publisher = publisher
-        book.genre = genre
-        book.image = image
-        tags = request.POST.getlist('tags')
-        book.tags.set(tags)
+            book.title = request.POST['title']
+            book.author = request.POST['author']
+            book.year = request.POST['year']
+            book.raiting = request.POST['raiting']
+            book.publisher = publisher
+            book.genre = genre
+            book.image = image
+            tags = request.POST.getlist('tags')
+            book.tags.set(tags)
 
-        book.save()
+            book.save()
 
-        return redirect("get_book", id=book.id)
+            return redirect("get_book", id=book.id)
+
+
